added assignment3-5 to folder chapter 3

+++ chapter 3 assignments/assignment3-3.cpp
@@ -4,8 +4,8 @@ using namespace std;
 
 int main()
 {
-  double celsius;
-  double fahrenhiet;
+  float celsius;
+  float fahrenhiet;
 
   cout << "Enter the tempreature in Celsius" << endl;
   cin >> celsius;
